Added the STL platform with the FileSystem implementation. Created the SDL Shell implementation which includes windows, clipboard, etc. Changed some platform functions for consistency (or shared their implementation generically).

+++ Extensions/Editor/ContentPackage.hpp
@@ -18,7 +18,7 @@ struct ContentPackageEntry
   // Name of the content file
   String File;
   // Size of the content item
-  uint Size;
+  u64 Size;
   bool Conflicted;
 };
